+++ src/main/java/com/project/online_book_store/app/service/OrdersService.java
@@ -9,5 +9,5 @@ import java.util.Map;
 public interface OrdersService {
     void createOrder(Long cartId);
 
-    Map<Integer, List<BuyBook>> createMapListOrders();
+    Map<Long, List<BuyBook>> createMapListOrders();
 }
